moved residual calculators and helpers to solver_residuals to reduce solvers file length
